backend/pyproject.toml
[project]
name = "aimops-backend"
version = "0.1.0"
description = "Artifact-centric image operations API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pillow>=10.4.0",
  "numpy>=2.0.0",
  "pyyaml>=6.0.2",
  "python-multipart>=0.0.9",
  "transformers>=4.44.0",
  "torch>=2.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
